+++ packages/headless/src/features/document-suggestion/document-suggestion-actions.ts
@@ -6,6 +6,7 @@ import {AsyncThunkCaseAssistOptions} from '../../api/service/case-assist/case-as
 import {prepareContextFromFields} from '../../api/service/case-assist/case-assist-params.js';
 import {GetDocumentSuggestionsRequest} from '../../api/service/case-assist/get-document-suggestions/get-document-suggestions-request.js';
 import {GetDocumentSuggestionsResponse} from '../../api/service/case-assist/get-document-suggestions/get-document-suggestions-response.js';
+import {NavigatorContext} from '../../app/navigatorContextProvider.js';
 import {
   CaseAssistConfigurationSection,
   DocumentSuggestionSection,
@@ -14,6 +15,8 @@ import {
   CaseInputSection,
   CaseFieldSection,
 } from '../../state/state-sections.js';
+import {SearchPageEvents} from '../analytics/search-action-cause.js';
+import {fromAnalyticsStateToAnalyticsParams} from '../configuration/analytics-params.js';
 
 export interface FetchDocumentSuggestionsThunkReturn {
   /** The successful document suggestions response. */
@@ -33,11 +36,14 @@ export const fetchDocumentSuggestions = createAsyncThunk<
   AsyncThunkCaseAssistOptions<StateNeededByFetchDocumentSuggestions>
 >(
   'caseAssist/documentSuggestions/fetch',
-  async (_, {getState, rejectWithValue, extra: {apiClient}}) => {
+  async (
+    _,
+    {getState, rejectWithValue, extra: {apiClient, navigatorContext}}
+  ) => {
     const state = getState();
 
     const fetched = await apiClient.getDocumentSuggestions(
-      await buildFetchDocumentSuggestionsRequest(state)
+      await buildFetchDocumentSuggestionsRequest(state, navigatorContext)
     );
 
     if (isErrorResponse(fetched)) {
@@ -51,7 +57,8 @@ export const fetchDocumentSuggestions = createAsyncThunk<
 );
 
 export const buildFetchDocumentSuggestionsRequest = async (
-  state: StateNeededByFetchDocumentSuggestions
+  state: StateNeededByFetchDocumentSuggestions,
+  navigatorContext: NavigatorContext
 ): Promise<GetDocumentSuggestionsRequest> => ({
   accessToken: state.configuration.accessToken,
   organizationId: state.configuration.organizationId,
@@ -65,6 +72,14 @@ export const buildFetchDocumentSuggestionsRequest = async (
   ...(state.configuration.analytics.enabled && {
     clientId: await getVisitorID(state.configuration.analytics),
   }),
+  ...(state.configuration.analytics.enabled &&
+    fromAnalyticsStateToAnalyticsParams(
+      state.configuration.analytics,
+      navigatorContext,
+      {
+        actionCause: SearchPageEvents.documentSuggestion,
+      }
+    )),
   fields: state.caseInput,
   context: state.caseField
     ? prepareContextFromFields(state.caseField.fields)